Updated NotificationController

+++ transaction-service/src/main/java/com/venuex/transaction_service/controller/NotificationController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.venuex.transaction_service.DTO.NotificationDTO;
 import com.venuex.transaction_service.service.NotificationService;
@@ -23,14 +24,47 @@ public class NotificationController {
     @GetMapping("/user/notifications")
     @ResponseStatus(HttpStatus.OK)
     public List<NotificationDTO> getAllNotifications(HttpServletRequest request) {
-        Integer userId = (Integer) request.getAttribute("userId");
+        String userIdHeader = request.getHeader("X-User-Id");
+
+        if (userIdHeader == null) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "MISING X-User-Id header");
+        }
+
+        Integer userId;
+
+        try {
+            userId = Integer.parseInt(userIdHeader);
+        } catch (NumberFormatException ex) {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Invalid X-User-Id format");
+        }
+
         return notificationService.getUserNotifications(userId);
     }
 
     @DeleteMapping("/user/notifications/{id}")
-    @ResponseStatus(HttpStatus.OK)
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteNotification(@PathVariable Integer id, HttpServletRequest request) {
-        Integer userId = (Integer) request.getAttribute("userId");
+        String userIdHeader = request.getHeader("X-User-ID");
+
+        if (userIdHeader == null) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "MISING X-User-Id header");
+        }
+
+        Integer userId;
+
+        try {
+            userId = Integer.parseInt(userIdHeader);
+        } catch (NumberFormatException ex) {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Invalid X-User-Id format");
+        }
         notificationService.deleteNotification(id, userId);
     }
 }
